fix: change node position payload type from int32 to float64 for improved precision

+++ internal/form/workflow/queries.sql
@@ -63,8 +63,8 @@ new_node AS (
         'label', 'New ' || initcap(@type::text),
         'type', @type::node_type,
         'payload', jsonb_build_object(
-            'x', COALESCE(@payload_x::int, 0),
-            'y', COALESCE(@payload_y::int, 0)
+            'x', COALESCE(@payload_x::float, 0),
+            'y', COALESCE(@payload_y::float, 0)
         )
     ) AS node
     FROM node_id
